% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{FKRCRT}
\alias{FKRCRT}
\title{Creates a Random Forest ensemble of k-Random Clusters Regression Trees (Fk-RCRT).}
\usage{
FKRCRT(x, y, n.trees = 100, mtry = NULL, k = 3, max.depth = Inf,
  var.improvp = 0.01, min.size = NULL, parallel = FALSE)
}
\arguments{
\item{x, }{y The input features and target variables respectively}

\item{n.trees}{The number of tree predictors (Default = 100)}

\item{mtry}{Feature subsample to build each tree. (Default = , as in CLUS)}

\item{k}{The number of random clusters to be generated at each split (Default = 3)}

\item{max.depth}{Maximum depth for generated trees (Default = Inf, split will are made while it is possible)}

\item{var.improvp}{Minimum variance decrease percentual when comparing a child to its parent needed to continue splitting (Default = 0.01)}

\item{min.size}{Minimum size of generated clusteres (Default = 5, as in CLUS)}

\item{parallel}{Whether to build the forest in parallel, using all disponible cores (Default = FALSE).}
}
\value{
A Fk-RCRT model
}
\description{
Creates a Random Forest ensemble of k-Random Clusters Regression Trees (Fk-RCRT).
}
