% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRCRT.R
\name{KRCRT}
\alias{KRCRT}
\title{Creates a k-Random Clusters Regression Trees (k-RCRTRF) model.}
\usage{
KRCRT(X, Y, k = 2, max.depth = Inf, var.improvp = 0.01, min.size = NULL)
}
\arguments{
\item{X, }{Y The input features and target variables respectively}

\item{k}{The number of random clusters to be generated at each split (Default = 3)}

\item{max.depth}{Maximum depth for generated trees (Default = Inf, split will are made while it is possible)}

\item{var.improvp}{Minimum variance decrease percentual when comparing a child to its parent needed to continue splitting (Default = 0.01)}

\item{min.size}{Minimum size of generated clusteres (Default = 5, as in CLUS)}
}
\value{
A k-RCRT model
}
\description{
Creates a k-Random Clusters Regression Trees (k-RCRTRF) model.
}
