% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{MORF}
\alias{MORF}
\title{Creates a Random Forest ensemble of Multi-target Regression Trees -- Multi-output Random Forest (MORF) (as in CLUS).}
\usage{
MORF(x, y, n.trees = 10, mtry = NULL, ftest.signf = 0.05, min.size = 5,
  max.depth = Inf, parallel = FALSE)
}
\arguments{
\item{x, }{y The input features and target variables respectively.}

\item{n.trees}{The number of tree predictors (Default = 100).}

\item{mtry}{Feature subsample to build each tree. (Default = , as in CLUS).}

\item{ftest.signf}{The signficance level for F-test's stopping criteria (Default = 0.05).}

\item{min.size}{Minimum size of generated clusteres (Default = 5, as in CLUS).}

\item{max.depth}{Maximum depth for generated trees (Default = Inf, split are made while it is possible).}

\item{parallel}{Whether to build the forest in parallel, using all disponible cores (Default = FALSE).}
}
\value{
A MORF model
}
\description{
Creates a Random Forest ensemble of Multi-target Regression Trees -- Multi-output Random Forest (MORF) (as in CLUS).
}
