% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRT.R
\name{MTRT}
\alias{MTRT}
\title{Creates Multi-target Regression Trees (MTRT), as in CLUS}
\usage{
MTRT(X, Y, ftest.signf = 0.05, min.size = 5, max.depth = Inf)
}
\arguments{
\item{X, }{Y The input features and target variables respectively}

\item{ftest.signf}{The signficance level for F-test's stopping criteria (Default = 0.05)}

\item{min.size}{Minimum size of generated clusteres (Default = 5, as in CLUS)}

\item{max.depth}{Maximum depth for generated trees (Default = Inf, split are made while it is possible)}
}
\value{
A MTRT model
}
\description{
Creates Multi-target Regression Trees (MTRT), as in CLUS
}
