% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{predict}
\alias{predict}
\title{Prediction function for Cluster-based Tree models.}
\usage{
predict(model, new.data, parallel = FALSE)
}
\arguments{
\item{model}{A MTRT, k-RCRT, MORF or k-RCRTRF model}

\item{new.data}{New instances to be predicted}
}
\value{
parallel A boolean indicating whether to predict data in parallel (Only applicable for ensemble models).
}
\description{
Prediction function for Cluster-based Tree models.
}
